*River of Gold*
I played a little piano during college. In my junior year, I officially learnt Ukulele as my first musical instrument. At that time, I wanted to improve my melody recognition, and someone that has so much knowledge and love for music like me should not end up not playing any instrument. While really love Ukulele, soon I realized that it is not suitable for slow and sad melodies but more for the happy and swing tempo. But the music genre I like is usually sensible, sad and painful. Guitar is similar in technics, but gives a deeper and more beautiful sound. So that is how I started to play guitar. 
I love guitar, but I disagree with the sayings that Uku is just a toy rather than an instrument. Uku is of its own advantages!